requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
accelerate=0.21.0=pypi_0
appdirs=1.4.4=pypi_0
asttokens=2.2.1=pypi_0
attrs=23.1.0=pypi_0
backcall=0.2.0=pypi_0
backports-functools-lru-cache=1.6.4=pypi_0
blas=1.0=mkl
brotlipy=0.7.0=py310h7f8727e_1002
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.05.30=h06a4308_0
cachetools=5.3.0=pypi_0
certifi=2022.12.7=pypi_0
cffi=1.15.1=py310h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=pypi_0
cmake=3.27.2=pypi_0
configargparse=1.5.3=pypi_0
contourpy=1.0.7=pypi_0
cryptography=38.0.4=pypi_0
cuda-cudart=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cycler=0.11.0=pypi_0
debugpy=1.5.1=pypi_0
decorator=5.1.1=pypi_0
diffusers=0.17.0=pypi_0
docker-pycreds=0.4.0=pypi_0
einops=0.6.0=pypi_0
entrypoints=0.4=pypi_0
executing=1.2.0=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.11.0=pypi_0
fire=0.5.0=pypi_0
flatbuffers=23.5.26=pypi_0
flit-core=3.6.0=pypi_0
fonttools=4.38.0=pypi_0
freetype=2.12.1=h4a9f257_0
fsspec=2023.6.0=pypi_0
ftfy=6.1.1=pypi_0
giflib=5.2.1=h5eee18b_3
gitdb=4.0.10=pypi_0
gitpython=3.1.31=pypi_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.16.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.51.1=pypi_0
huggingface-hub=0.16.4=pypi_0
idna=3.4=py310h06a4308_0
imageio=2.25.0=pypi_0
imageio-ffmpeg=0.4.8=pypi_0
importlib-metadata=6.0.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46305
ipykernel=6.15.0=pypi_0
ipython=8.12.0=pypi_0
jedi=0.18.2=pypi_0
jinja2=3.1.2=pypi_0
jpeg=9e=h5eee18b_1
jsonschema=4.19.1=pypi_0
jsonschema-specifications=2023.7.1=pypi_0
jupyter-client=7.0.6=pypi_0
jupyter-core=4.12.0=pypi_0
kiwisolver=1.4.4=pypi_0
kornia=0.6.9=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufile=1.7.0.149=0
libcurand=10.3.3.53=0
libcusolver=11.4.0.1=0
libcusparse=11.7.4.91=0
libdeflate=1.17=h5eee18b_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libnpp=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.0=h6a678d5_2
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
lit=16.0.6=pypi_0
loguru=0.7.0=pypi_0
lpips=0.1.4=pypi_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.1=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.6.3=pypi_0
matplotlib-inline=0.1.6=pypi_0
mediapipe=0.10.1=pypi_0
mkl=2023.1.0=h6d00ec8_46342
mkl-service=2.4.0=py310h5eee18b_1
mkl_fft=1.3.6=py310h1128e8f_1
mkl_random=1.2.2=py310h1128e8f_1
mpmath=1.3.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=pypi_0
nettle=3.7.3=hbbd107a_1
networkx=3.0=pypi_0
numpy=1.23.5=pypi_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-cupti-cu11=11.7.101=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-curand-cu11=10.2.10.91=pypi_0
nvidia-cusolver-cu11=11.4.0.1=pypi_0
nvidia-cusparse-cu11=11.7.4.91=pypi_0
nvidia-nccl-cu11=2.14.3=pypi_0
nvidia-nvtx-cu11=11.7.91=pypi_0
oauthlib=3.2.2=pypi_0
opencv-contrib-python=4.8.0.74=pypi_0
opencv-python-headless=4.8.0.74=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=3.0.9=h7f8727e_0
packaging=23.0=pypi_0
parso=0.8.3=pypi_0
pathtools=0.1.2=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.3.0=pypi_0
pip=23.2.1=pypi_0
plyfile=0.7.4=pypi_0
prompt-toolkit=3.0.38=pypi_0
protobuf=3.20.3=pypi_0
psutil=5.9.4=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.15.0=pypi_0
pyopenssl=22.0.0=pypi_0
pyparsing=3.0.9=pypi_0
pyproject=1.3.1=pypi_0
pyproject-toml=0.0.10=pypi_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.12=h955ad1f_0
python-dateutil=2.8.2=pypi_0
pytorch-cuda=11.7=h778d358_5
pytorch-mutex=1.0=cuda
pywavelets=1.4.1=pypi_0
pyyaml=6.0=pypi_0
pyzmq=19.0.2=pypi_0
readline=8.2=h5eee18b_0
referencing=0.30.2=pypi_0
regex=2023.3.23=pypi_0
requests=2.28.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rpds-py=0.10.3=pypi_0
rsa=4.9=pypi_0
safetensors=0.3.1=pypi_0
scikit-image=0.19.3=pypi_0
scipy=1.10.0=pypi_0
sentry-sdk=1.15.0=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=65.6.3=pypi_0
six=1.16.0=pypi_0
smmap=5.0.0=pypi_0
sounddevice=0.4.6=pypi_0
sqlite=3.41.2=h5eee18b_0
stack-data=0.6.2=pypi_0
sympy=1.12=pypi_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.11.2=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
termcolor=2.3.0=pypi_0
tifffile=2023.2.3=pypi_0
timm=0.9.2=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.3=pypi_0
toml=0.10.2=pypi_0
torch=2.0.1=pypi_0
torchaudio=0.13.1=py310_cu117
torchmetrics=0.11.4=pypi_0
torchsummary=1.5.1=pypi_0
torchvision=0.14.1=py310_cu117
tornado=6.1=pypi_0
tqdm=4.64.1=pypi_0
traitlets=5.9.0=pypi_0
transformers=4.31.0=pypi_0
triton=2.0.0=pypi_0
typing-extensions=4.4.0=pypi_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.14=pypi_0
wandb=0.13.10=pypi_0
wcwidth=0.2.6=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pypi_0
xformers=0.0.21=pypi_0
xz=5.4.2=h5eee18b_0
zipp=3.12.1=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
